Code cleanup in Lowest Common Ancestor
